% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import one month of data}
\usage{
import(Month, removeNA = TRUE)
}
\arguments{
\item{Month}{File handle for raw data, excluding the \verb{data/} and \code{.txt} extension (character)}

\item{removeNA}{Remove any rows with missing data? Default: TRUE. (logical)}
}
\value{
A \code{data.table}
}
\description{
Import one month of data
}
\examples{
May <- import('may')
head(May)
}
