% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month-merge.R
\name{month_merge}
\alias{month_merge}
\title{Import and merge two months of data}
\usage{
month_merge(x, y, ...)
}
\arguments{
\item{x, y}{The \code{Month} arguments to \code{import()}, which is called twice (character)}

\item{...}{Additional arguments to \code{import()}}
}
\value{
A single \code{data.table} with two months of data
}
\description{
Import and merge two months of data
}
\examples{
DT <- month_merge('may', 'june')
print(DT)
}
